♻️ [Refactor]: modal자체에서 isOpen 관리
- globalModal이 관리하던 isOpen state를 자체 modal에 선언
- ModalContainer 분리


Bug: #9

<Issue Tracker>
Fixes: #9
