+++ src/components/organisms/modals/ConfirmModal.tsx
@@ -4,9 +4,11 @@ import Button from '@/components/atoms/Button';
 import Container from '@/components/atoms/Container';
 import Typography from '@/components/atoms/Typography';
 import { ConfirmModalAtom } from '@/stores/globalModal.store';
+import ModalContainer from '@/components/organisms/modals/ModalContainer';
 
 export default function ConfirmModal() {
   const {
+    isOpen,
     title,
     message,
     confirmButtonContent,
@@ -15,8 +17,9 @@ export default function ConfirmModal() {
     onClickCancel,
   } = useRecoilValue(ConfirmModalAtom);
 
-  return (
-    <Container.FlexRow className="fixed left-0 top-0 z-50 h-[100vh] w-[100vw] items-center justify-center bg-[#6D6D6D]/50">
+  console.log('confirm modal rendered!!');
+  return isOpen ? (
+    <ModalContainer>
       <Container.FlexCol className="w-full max-w-96 gap-3 rounded-2xl bg-bg p-6 text-brown">
         <Container.FlexCol className="gap-6">
           <Typography.SubTitle3>{title}</Typography.SubTitle3>
@@ -43,8 +46,8 @@ export default function ConfirmModal() {
           </Button.Ghost>
         </Container.FlexRow>
       </Container.FlexCol>
-    </Container.FlexRow>
-  );
+    </ModalContainer>
+  ) : null;
 }
 
 ConfirmModal.defaultProps = {
