+++ src/components/organisms/modals/ProfileModal.tsx
@@ -1,9 +1,14 @@
 import { useRecoilValue } from 'recoil';
 
 import { ProfileModalAtom } from '@/stores/globalModal.store';
+import ModalContainer from '@/components/organisms/modals/ModalContainer';
 
 export default function ProfileModal() {
-  const profileModalState = useRecoilValue(ProfileModalAtom);
+  const { isOpen } = useRecoilValue(ProfileModalAtom);
 
-  return <div>ConfirmModal</div>;
+  return isOpen ? (
+    <ModalContainer>
+      <div>ConfirmModal</div>
+    </ModalContainer>
+  ) : null;
 }
